feat: update layouts
