+++ app/(app)/(forAuthorizedUsers)/(withMenu)/layout.tsx
@@ -1,0 +1,14 @@
+'use client'
+
+import { ReactNode } from 'react'
+
+import { MenuApp } from '@/widgets'
+
+export default function WithMenuLayout({ children }: { children: ReactNode }) {
+  return (
+    <>
+      {children}
+      <MenuApp />
+    </>
+  )
+}
