+++ app/(app)/(forAuthorizedUsers)/(withoutMenu)/layout.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import { ReactNode } from 'react'
+
+export default function WithoutMenuLayout({ children }: { children: ReactNode }) {
+  return <></>
+}
