--- app/(public)/layout.tsx
@@ -1,14 +1,0 @@
-'use client'
-
-import { ReactNode } from 'react'
-
-import { HeaderPublic } from '@/widgets'
-
-export default function PublicLayout({ children }: { children: ReactNode }) {
-  return (
-    <>
-      <HeaderPublic />
-      <main>{children}</main>
-    </>
-  )
-}
