+++ features/publicPage/ui/PublicPage.module.scss
@@ -3,6 +3,4 @@
   flex-direction: column;
   gap: 36px;
   align-items: center;
-
-  padding: 24px;
 }
